share test fixtures between sqlite and indexedDB
