+++ packages/store/src/database/backends/sqlite.test.ts
@@ -1,7 +1,6 @@
-import { EventID, Task, TaskID } from "../../core2";
+import { EventID, TaskID } from "../../core2";
 import { EntityType } from "../../core2/entities/entityBase";
-import { testTask } from "../__tests__/testTasks";
-import { DatabaseBackendEntityRecord } from "../backend";
+import { createTestTask, testTasks } from "../__tests__/testTasks";
 import {
   constructByIDSQLQuery,
   constructEntitySQLQuery,
@@ -28,28 +27,6 @@ test("DB automatically migrates", async () => {
   const db = await backend.__accessDBForTesting();
   expect(await getSchemaVersionNumber(db)).toBe(latestSchemaVersionNumber);
 });
-
-function createTestTask({
-  id,
-  lastEventID,
-  dueTimestampMillis,
-}: {
-  id: string;
-  lastEventID: string;
-  dueTimestampMillis: number;
-}): DatabaseBackendEntityRecord<Task> {
-  // lazy deep clone
-  const newTask = JSON.parse(JSON.stringify(testTask)) as Task;
-  newTask.id = id as TaskID;
-  newTask.componentStates["a"].dueTimestampMillis = dueTimestampMillis;
-  return { entity: newTask, lastEventID: lastEventID as EventID };
-}
-
-const testTasks: DatabaseBackendEntityRecord<Task>[] = [
-  createTestTask({ id: "a", lastEventID: "x", dueTimestampMillis: 50 }),
-  createTestTask({ id: "b", lastEventID: "y", dueTimestampMillis: 100 }),
-  createTestTask({ id: "c", lastEventID: "z", dueTimestampMillis: 150 }),
-];
 
 test("round-trip entities", async () => {
   await backend.putEntities(testTasks);
